+++ packages/palette/src/elements/index.tsx
@@ -21,7 +21,6 @@ export * from "./ReadMore"
 export * from "./Select"
 export * from "./Separator"
 export * from "./Slider"
-export * from "./Spacer"
 export * from "./Spinner"
 export * from "./StackableBorderBox"
 export * from "./Stepper"
